backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
assemblyai==0.17.0
python-jose==3.3.0
gunicorn==21.2.0